requirements.txt
py-cpuinfo==9.0.0
loguru==0.7.0
kubernetes==26.1.0
